functions/download-handler/requirements.txt
google-cloud-iot~=1.0.0
google-cloud-storage~=1.30.0
requests~=2.21.0
google-auth~=1.20.1
google-api-core~=1.22.1